+++ src/backend/ci/core/auth/biz-auth-blueking/build.gradle
@@ -28,7 +28,7 @@ dependencies {
     compile project(":core:common:common-service")
     compile project(":core:common:common-client")
     compile project(":core:common:common-auth:common-auth-api")
-    compile project(":core:project:api-project")
+    compile project(":core:project:api-project-sample")
     compile project(":core:process:api-process")
     compile project(":core:repository:api-repository")
     compile project(":core:environment:api-environment")
