+++ src/backend/ci/core/environment/api-environment/src/main/kotlin/com/tencent/devops/environment/api/ServiceNodeResource.kt
@@ -131,19 +131,4 @@ interface ServiceNodeResource {
         @PathParam("nodeType")
         nodeType: NodeType
     ): Result<List<NodeBaseInfo>>
-
-    @ApiOperation("分页获取节点列表")
-    @GET
-    @Path("/projects/{projectId}/list/page")
-    fun listNodeByPage(
-        @ApiParam("项目ID", required = true)
-        @PathParam("projectId")
-        projectId: String,
-        @ApiParam("第几页", required = false, defaultValue = "1")
-        @QueryParam("page")
-        page: Int? = null,
-        @ApiParam("每页多少条", required = false, defaultValue = "20")
-        @QueryParam("pageSize")
-        pageSize: Int? = null
-    ): Result<Page<NodeBaseInfo>>
 }
