+++ src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/api/service/ServicePipelineResource.kt
@@ -300,22 +300,4 @@ interface ServicePipelineResource {
         @PathParam("pipelineId")
         pipelineId: String
     ): Result<Boolean>
-
-    @ApiOperation("流水线编排列表")
-    @GET
-    @Path("/{projectId}/list")
-    fun pipelineList(
-        @ApiParam("项目ID", required = true)
-        @PathParam("projectId")
-        projectId: String,
-        @ApiParam("第几页", required = false, defaultValue = "1")
-        @QueryParam("page")
-        page: Int? = null,
-        @ApiParam("每页多少条", required = false, defaultValue = "20")
-        @QueryParam("pageSize")
-        pageSize: Int? = null,
-        @ApiParam("渠道号，默认为DS", required = false)
-        @QueryParam("channelCode")
-        channelCode: ChannelCode? = ChannelCode.BS
-    ): Result<PipelineViewPipelinePage<PipelineInfo>>
 }
