+++ src/backend/ci/core/project/biz-project/src/main/kotlin/com/tencent/devops/project/resources/ServiceProjectResourceImpl.kt
@@ -100,7 +100,4 @@ class ServiceProjectResourceImpl @Autowired constructor(
         return Result(projectService.update(userId, projectId, projectUpdateInfo))
     }
 
-    override fun list(limit: Int, offset: Int): Result<Page<ProjectVO>> {
-        return Result(projectService.list(limit, offset))
-    }
 }
