+++ src/backend/ci/gradle.properties
@@ -32,7 +32,7 @@ org.gradle.daemon.idleTimeout=3600000
 #systemProp.http.proxyPort=
 #systemProp.https.proxyHost=
 #systemProp.https.proxyPort=
-MAVEN_REPO_URL=http://mirrors.tencent.com/nexus/repository/maven-public
+MAVEN_REPO_URL=http://mirrors.cloud.tencent.com/nexus/repository/maven-public
 MAVEN_REPO_DEPLOY_URL=
 MAVEN_REPO_USERNAME=
 MAVEN_REPO_PASSWORD=
