+++ src/backend/ci/task_spring_boot_package.gradle
@@ -31,7 +31,6 @@ apply plugin: "kotlin"
 apply plugin: "kotlin-spring"
 apply plugin: "org.springframework.boot"
 
-
 dependencies {
     compile project(":core:common:common-api")
     compile project(":core:common:common-web")
