--- support-files/iam/0005_bk_ci_20200623-2120_iam.json
@@ -1,35 +1,0 @@
-{
-  "system_id": "bk_ci",
-  "operations": [
-    {
-      "operation": "delete_action",
-      "data":{
-        "id": "env_node_view"
-      }
-    },
-    {
-      "operation": "delete_action",
-      "data": {
-        "id": "env_node_edit"
-      }
-    },
-    {
-      "operation": "delete_action",
-      "data": {
-        "id": "env_node_create"
-      }
-    },
-    {
-      "operation": "delete_action",
-      "data": {
-        "id": "env_node_delete"
-      }
-    },
-    {
-      "operation": "delete_action",
-      "data": {
-        "id": "env_node_use"
-      }
-    }
-  ]
-}
